[FEAT] 최근 감정 기록 조회 기능 #31

+++ src/main/java/goormthonuniv/team_22_be/questionanswer/application/dto/DailyProgressResponse.java
@@ -3,18 +3,18 @@ package goormthonuniv.team_22_be.questionanswer.application.dto;
 public record DailyProgressResponse(
 
         Long answerQuestions,
-        Double percentage
+        Long percentage
 ) {
 
     private static final int TOTAL_GOAL_COUNT = 6;
 
     public static DailyProgressResponse from(Long answerQuestions) {
         if (answerQuestions == null || answerQuestions <= 0) {
-            return new DailyProgressResponse(0L, 0.0);
+            return new DailyProgressResponse(0L, 0L);
         }
 
         double calculatedPercentage = ((double) answerQuestions / TOTAL_GOAL_COUNT) * 100;
 
-        return new DailyProgressResponse(answerQuestions, calculatedPercentage);
+        return new DailyProgressResponse(answerQuestions, (long) calculatedPercentage);
     }
 }
